Find-icament/requirements.txt
bs4
mysql-connector-python
lxml
requests